quikappappleassistant/scripts/requirements.txt
pyopenssl>=23.0.0
cryptography>=41.0.0 